combined keymap and keyset IterOption in a different way
